+++ src/Typin/Typin.Core/Pipeline.cs
@@ -1,0 +1,10 @@
+﻿namespace Typin
+{
+    using System.Threading.Tasks;
+
+    /// <summary>
+    /// Represents an async continuation for the next task to execute in the pipeline.
+    /// </summary>
+    /// <returns>Awaitable task</returns>
+    public delegate Task CommandPipelineHandlerDelegate();
+}