+++ src/Typin/Typin.Tests/Data/CustomDirectives/Invalid/AbstractDirective.cs
@@ -1,21 +1,14 @@
 ﻿namespace Typin.Tests.Data.CustomDirectives.Invalid
 {
+    using System.Threading;
     using System.Threading.Tasks;
     using Typin;
     using Typin.Attributes;
-    using Typin.Console;
 
     [Directive("invalid-abstract", Description = "Abstract directive.")]
     public abstract class AbstractDirective : IDirective
     {
-        public bool ContinueExecution => true;
-
-        public AbstractDirective()
-        {
-
-        }
-
-        public ValueTask HandleAsync(IConsole console)
+        public ValueTask OnInitializedAsync(CancellationToken cancellationToken)
         {
             return default;
         }