+++ src/Typin/Typin.Tests/Data/CustomDirectives/Valid/CustomDirective.cs
@@ -1,27 +1,25 @@
 ﻿namespace Typin.Tests.Data.CustomDirectives.Valid
 {
+    using System.Threading;
     using System.Threading.Tasks;
     using Typin;
     using Typin.Attributes;
-    using Typin.Console;
 
     [Directive("custom", Description = "Custom directive.")]
     public sealed class CustomDirective : IDirective
     {
         public const string ExpectedOutput = nameof(CustomDirective);
 
-        public bool ContinueExecution => true;
-
-        public CustomDirective()
+        public ValueTask OnInitializedAsync(CancellationToken cancellationToken)
         {
-
+            return default;
         }
 
-        public ValueTask HandleAsync(IConsole console)
+        public async ValueTask HandleAsync(ICliContext context, CommandPipelineHandlerDelegate next, CancellationToken _)
         {
-            console.Output.Write(ExpectedOutput);
+            context.Console.Output.Write(ExpectedOutput);
 
-            return default;
+            await next();
         }
     }
 }