+++ src/Typin/Typin.Tests/Data/CustomDirectives/Valid/CustomThrowableDirectiveWithInnerException.cs
@@ -1,10 +1,10 @@
 ﻿namespace Typin.Tests.Data.CustomDirectives.Valid
 {
     using System;
+    using System.Threading;
     using System.Threading.Tasks;
     using Typin;
     using Typin.Attributes;
-    using Typin.Console;
     using Typin.Exceptions;
 
     [Directive("custom-throwable-with-inner-exception", Description = "Custom throwable directive with message.")]
@@ -14,16 +14,14 @@
         public const string ExpectedExceptionMessage = nameof(CustomThrowableDirectiveWithInnerException) + "ExMessage";
         public const int ExpectedExitCode = 2;
 
-        public bool ContinueExecution => true;
-
-        public CustomThrowableDirectiveWithInnerException()
+        public ValueTask OnInitializedAsync(CancellationToken cancellationToken)
         {
-
+            return default;
         }
 
-        public ValueTask HandleAsync(IConsole console)
+        public ValueTask HandleAsync(ICliContext context, CommandPipelineHandlerDelegate next, CancellationToken _)
         {
-            console.Output.Write(ExpectedOutput);
+            context.Console.Output.Write(ExpectedOutput);
 
             throw new DirectiveException(ExpectedExceptionMessage, new NullReferenceException(), ExpectedExitCode);
         }