+++ src/Typin/Typin.Tests/Data/CustomDirectives/Valid/CustomThrowableDirectiveWithMessageAndShowHelp.cs
@@ -1,9 +1,9 @@
 ﻿namespace Typin.Tests.Data.CustomDirectives.Valid
 {
+    using System.Threading;
     using System.Threading.Tasks;
     using Typin;
     using Typin.Attributes;
-    using Typin.Console;
     using Typin.Exceptions;
 
     [Directive("custom-throwable-with-message-and-show-help", Description = "Custom throwable directive with message and show help.")]
@@ -13,16 +13,14 @@
         public const string ExpectedExceptionMessage = nameof(CustomThrowableDirectiveWithMessageAndShowHelp) + "ExMessage";
         public const int ExpectedExitCode = 2;
 
-        public bool ContinueExecution => true;
-
-        public CustomThrowableDirectiveWithMessageAndShowHelp()
+        public ValueTask OnInitializedAsync(CancellationToken cancellationToken)
         {
-
+            return default;
         }
 
-        public ValueTask HandleAsync(IConsole console)
+        public ValueTask HandleAsync(ICliContext context, CommandPipelineHandlerDelegate next, CancellationToken _)
         {
-            console.Output.Write(ExpectedOutput);
+            context.Console.Output.Write(ExpectedOutput);
 
             throw new DirectiveException(ExpectedExceptionMessage, ExpectedExitCode, true);
         }