+++ src/Typin/Typin/Directives/DebugDirective.cs
@@ -3,6 +3,7 @@
     using System;
     using System.Diagnostics;
     using System.Diagnostics.CodeAnalysis;
+    using System.Threading;
     using System.Threading.Tasks;
     using Typin.Attributes;
     using Typin.Console;
@@ -13,15 +14,19 @@
     /// </summary>
     [ExcludeFromCodeCoverage]
     [Directive(BuiltInDirectives.Debug, Description = "Starts a debugging mode. Application will wait for debugger to be attached before proceeding.")]
-    public sealed class DebugDirective : IDirective
+    public sealed class DebugDirective : IPipelinedDirective
     {
         /// <inheritdoc/>
-        public bool ContinueExecution => true;
+        public ValueTask OnInitializedAsync(CancellationToken cancellationToken)
+        {
+            return default;
+        }
 
         /// <inheritdoc/>
-        public async ValueTask HandleAsync(IConsole console)
+        public async ValueTask HandleAsync(ICliContext context, CommandPipelineHandlerDelegate next, CancellationToken cancellationToken)
         {
             int processId = Process.GetCurrentProcess().Id;
+            IConsole console = context.Console;
 
             console.WithForegroundColor(ConsoleColor.Green, () =>
                 console.Output.WriteLine($"Attach debugger to PID {processId} to continue."));
@@ -29,11 +34,13 @@
             Debugger.Launch();
 
             while (!Debugger.IsAttached)
-                await Task.Delay(100);
+                await Task.Delay(100, cancellationToken);
 
             //Replace with an event
             console.WithForegroundColor(ConsoleColor.Green, () =>
                 console.Output.WriteLine($"Debugger attached to PID {processId}."));
+
+            await next();
         }
     }
 }