+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Merged `HandleVersionOption` and `HandleHelpOption` into one middleware named `HandleSpecialOptions`.
 - Core middleware execution order has changed: `ResolveCommandSchema` -> `ResolveCommandInstance` -> [User middlewares] -> `HandleSpecialOptions` -> `ExecuteCommand`).
 - Removed unnecessary casts to `CliContext` from `ICliContext`.
+- Removed `IDirective.ContinueExecution`, modified `IDirective`, and added `IPipelinedDirective`.
 
 ### v2.1.1 (18-Oct-2020)